continued working on sorting.

+++ database.py
@@ -18,7 +18,7 @@ def init_db():
         CREATE TABLE IF NOT EXISTS entries (
             id INTEGER PRIMARY KEY AUTOINCREMENT,
             author TEXT,
-            importance INTEGER,
+            importance REAL,
             title TEXT,
             content TEXT
         )
